requirements.txt
django==5.1.6
djangorestframework==3.15.2